+++ src/core/states.rs
@@ -1,10 +1,11 @@
 #![allow(dead_code)]
 
 use std::thread::JoinHandle;
+use std::net::SocketAddr;
 
 use super::config::ServerConfig;
 use super::router::Route;
-use crate::channel::{self, TryRecvError};
+use crate::channel::{self, TryRecvError, SendError};
 use crate::support::debug::{self, InfoLevel};
 use crate::support::session::*;
 
@@ -16,12 +17,36 @@ pub enum ControlMessage {
     Custom(String),
 }
 
+pub struct AsyncController(channel::Sender<ControlMessage>, SocketAddr);
+
+impl AsyncController {
+    fn new(messenger: channel::Sender<ControlMessage>, addr: SocketAddr) -> Self {
+        AsyncController(messenger, addr)
+    }
+
+    pub fn send(&self, message: ControlMessage) -> Result<(), SendError<ControlMessage>> {
+        self.0.send(message)?;
+
+        match message {
+            ControlMessage::Terminate => {
+                //TODO: connect to self address --
+                // let s = TcpStream::connect(self.1)
+                // if let Ok(ss) = s { let _ = ss.shutdown(Shutdown::Both); }
+            }
+            _ => {}
+        };
+
+        Ok(())
+    }
+}
+
 pub struct ServerStates {
     running: bool,
     courier_channel: (
         channel::Sender<ControlMessage>,
         channel::Receiver<ControlMessage>,
     ),
+    socket_addr: SocketAddr,
     session_auto_clean_handler: Option<JoinHandle<()>>,
 }
 
@@ -30,6 +55,7 @@ impl ServerStates {
         ServerStates {
             running: false,
             courier_channel: channel::bounded(1),
+            socket_addr: SocketAddr::from(([127, 0, 0, 1], 8080)),
             session_auto_clean_handler: None,
         }
     }
@@ -44,9 +70,13 @@ impl ServerStates {
         }
     }
 
+    pub(crate) fn set_port(&mut self, port: u16) {
+        self.socket_addr.set_port(port);
+    }
+
     #[inline]
-    pub(crate) fn get_courier_sender(&self) -> channel::Sender<ControlMessage> {
-        self.courier_channel.0.clone()
+    pub(crate) fn get_courier_sender(&self) -> AsyncController {
+        AsyncController::new(self.courier_channel.0.clone(), self.socket_addr)
     }
 
     pub(crate) fn courier_deliver(&self, msg: ControlMessage) -> Result<(), channel::SendError<ControlMessage>> {